lesson: Signing In and Out, add partial LS solution
The LS solution just had the index page changed so that if a user wasn't signed
in, they wouldn't see the content on the index page, but they didn't change the
layout to prevent other pages from being displayed.

I thought it would be better to hide all of the content so I changed the
layout.erb to not display any content unless a user is signed in.

To enable the signin.erb page to have the mising html elements that would have
been in the layout.erb, I added the elements directly to the signin.erb file,
and when calling erb in the route, I added a false argement for which layout to
use:

`erb signin, layout: false`